requirements_enhanced.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
hdbscan>=0.8.29
umap-learn>=0.5.3

# Azure dependencies
azure-storage-blob>=12.14.0
azure-identity>=1.12.0
openai>=1.0.0

# BigQuery dependencies
google-cloud-bigquery>=3.4.0
google-auth>=2.15.0
db-dtypes>=1.0.5

# Text processing
nltk>=3.8
spacy>=3.4.0
textstat>=0.7.3

# Utilities
PyYAML>=6.0
python-dotenv>=0.19.0
schedule>=1.2.0
joblib>=1.2.0
tqdm>=4.64.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Logging and monitoring
loguru>=0.6.0

# Optional: For notebook support
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0